fix build on win32

+++ src/xtd.forms/src/xtd/forms/vertical_layout_panel.cpp
@@ -42,7 +42,7 @@ void vertical_layout_panel::on_layout(const event_args& e) {
 
   int32_t absolute_height = 0;
   for (auto& [control, layout_style] : control_layout_styles_)
-    if (control.get().visible() && layout_style.size_type() == size_type::absolute) absolute_height += layout_style.height().value_or(control.get().default_size().height());
+    if (control.get().visible() && layout_style.size_type() == size_type::absolute) absolute_height += static_cast<int32_t>(layout_style.height().value_or(control.get().default_size().height()));
   auto_size_height -= absolute_height;
 
   float total_percent = 0;
@@ -64,9 +64,9 @@ void vertical_layout_panel::on_layout(const event_args& e) {
       else  left = client_size().width() / 2 - control.get().width() / 2;
     }
     width = layout_style.expanded() ? client_size().width() - padding().left() - padding().right() : control.get().width();
-    if (layout_style.size_type() == size_type::absolute) height = layout_style.height().value_or(control.get().default_size().height());
+    if (layout_style.size_type() == size_type::absolute) height = static_cast<int32_t>(layout_style.height().value_or(control.get().default_size().height()));
     else if (layout_style.size_type() == size_type::percent) height = static_cast<int32_t>(percent_height * (layout_style.height().value_or(0) / total_percent));
-    else if (layout_style.size_type() == size_type::auto_size) height = auto_size_height / auto_size_control_count;
+    else if (layout_style.size_type() == size_type::auto_size) height = static_cast<int32_t>(auto_size_height / auto_size_control_count);
     else throw argument_exception(caller_info_);
     control.get().set_bounds(left, top, width, height);
     top += height + padding().bottom() + padding().top();
